+++ internal/ecs/resource.go
@@ -27,6 +27,9 @@ func Registry(cl config.Cluster) []ResourceCollector {
 	}
 	if cl.MeteringEnabled() {
 		rcs = append(rcs, Metering{})
+		if cl.QuotasEnabled() {
+			rcs = append(rcs, Quotas{})
+		}
 	}
 	if cl.CollectDT {
 		rcs = append(rcs, NewDT(cl))